requirements.txt
aiogram==3.4.1
aiosqlite==0.20.0
flask==3.0.2
python-dotenv==1.0.1
